- New network to support unpacking to different prootbuf messages - Fixing errors in talkers behaviors - Fixing update errors in PBReceiver

+++ Talkers/Behaviours/T-3/B-18/O-18.hpp
@@ -32,7 +32,7 @@ public:
         */
 
         _state["SUDDEN_MOVE_SPEED"] = behaviourState["SUDDEN_MOVE_SPEED_DOUBLE"];
-		return std::get<double>(behaviourState["RAND"]) < 33;
+		return std::get<int>(behaviourState["RAND"]) < 33;
     }
 };
