+++ Talkers/Behaviours/T-3/B-18/O-19.hpp
@@ -30,7 +30,7 @@ public:
         Condition: 1/3 chance
         */
 
-        return std::get<double>(behaviourState["RAND"]) < 66;
+        return std::get<int>(behaviourState["RAND"]) < 66;
     }
 };
